requirements.txt
PyQt6>=6.0.0
Pillow>=8.0.0
orjson>=3.0.0
psutil>=5.8.0
nuitka>=1.8.0 